separate RescanProgress
