+++ app/components/indicators/index.js
@@ -7,7 +7,7 @@ export { default as NoMoreTicketsIndicator } from "./NoMoreTickets";
 export { default as NoTicketsIndicator } from "./NoTickets";
 export { default as DecredLoading } from "./DecredLoading/DecredLoading";
 export { default as SimpleLoading } from "./SimpleLoading";
-export { default as RescanProgress } from "./RescanProgress";
+export { default as RescanProgress } from "./RescanProgress/RescanProgress";
 export { default as AnimatedLinearProgressFull } from "./AnimatedLinearProgressFull";
 export { default as LinearProgressSmall } from "./LinearProgressSmall";
 export { default as LoaderBarBottom } from "./LoaderBarBottom";
